Send audio to mixer for single user in conference when recording.
Right now, if a user is alone in the conference, they never send any
audio to the mixer. This has an unfortunate side-effect that while a
conference with only one user can be recorded (if you start recording
manually), the recording will contain only silence.

Add an explicit check for recording, and allow that to override the
conference member count check.